user / auth 분리, user rest api 설정
