+++ trip-location/src/main/java/com/korea/triplocation/api/controller/AuthController.java
@@ -2,16 +2,19 @@ package com.korea.triplocation.api.controller;
 
 import javax.validation.Valid;
 
-import lombok.Getter;
 import org.springframework.http.ResponseEntity;
 import org.springframework.validation.BindingResult;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
 
 import com.korea.triplocation.aop.annotation.ValidAspect;
 import com.korea.triplocation.api.dto.request.LoginReqDto;
 import com.korea.triplocation.api.dto.request.UserReqDto;
 import com.korea.triplocation.api.dto.response.DataRespDto;
-import com.korea.triplocation.service.UserService;
+import com.korea.triplocation.service.AuthService;
 
 import lombok.RequiredArgsConstructor;
 
@@ -20,14 +23,14 @@ import lombok.RequiredArgsConstructor;
 @RequestMapping("/api/v1/auth")
 public class AuthController {
 	
-	private final UserService userService;
+	private final AuthService authService;
 	
 	@ValidAspect
     @PostMapping("/user")
     public ResponseEntity<?> signup(@Valid @RequestBody UserReqDto userReqDto, BindingResult bindingResult) {
 //    	System.out.println(userReqDto.getEmail() + "\t" + userReqDto.getPassword());
-    	userService.checkDuplicatedByEmail(userReqDto.getEmail());
-    	userService.signup(userReqDto);
+		authService.checkDuplicatedByEmail(userReqDto.getEmail());
+		authService.signup(userReqDto);
         return ResponseEntity.ok(DataRespDto.ofDefault());
     }
 	
@@ -37,16 +40,16 @@ public class AuthController {
 		System.out.println(loginReqDto.getEmail() + "\t" + loginReqDto.getPassword());
 		System.out.println(bindingResult.toString());
 //		userService.signin(loginReqDto);
-        return ResponseEntity.ok(userService.signin(loginReqDto));
+        return ResponseEntity.ok(authService.signin(loginReqDto));
     }
 
     @GetMapping("/authenticated")
     public ResponseEntity<?> authenticated(String accessToken) {
-        return ResponseEntity.ok(userService.authenticated(accessToken));
+        return ResponseEntity.ok(authService.authenticated(accessToken));
     }
 
     @GetMapping("/principal")
     public ResponseEntity<?> principal(String accessToken) {
-        return ResponseEntity.ok(userService.getPrincipal(accessToken));
+        return ResponseEntity.ok(authService.getPrincipal(accessToken));
     }
 }
