+++ trip-location/src/main/java/com/korea/triplocation/repository/UserRepository.java
@@ -1,14 +1,5 @@
 package com.korea.triplocation.repository;
 
-import org.apache.ibatis.annotations.Mapper;
+public class UserRepository {
 
-import com.korea.triplocation.domain.user.entity.Authority;
-import com.korea.triplocation.domain.user.entity.User;
-
-@Mapper
-public interface UserRepository {
-
-	public int saveUser(User user);
-	public int saveAuthority(Authority authority);
-	public User findUserByEmail(String email);
 }
